IOSS: Unify TextMesh unit tests in Ioss and STK
1) Created single unit test file that can be copied between systems
2) Created associated test fixtures that support the test
